New script for starting SmartDoorHost. New script for monitoring SmartDoorHost. New script for Updated configuration to contain python_version as we use it to start our host.

+++ KillSmartDoor.py
@@ -1,22 +1,31 @@
 import os
 import sys
 import signal
-import subprocess
+import MonitorSmartDoor
 
-process_count = 0
-p = subprocess.Popen(['pgrep', '-f', 'SmartDoorHost'], stdout=subprocess.PIPE)
-output, err = p.communicate()
 
-for line in output.splitlines():
-    process_count += 1
-    line = bytes.decode(line)
-    print("Killing SmartDoor process: {0}".format(line))
-    pid = int(line.split(None, 1)[0])
-    os.kill(pid, signal.SIGTERM)
+def kill():
+    """ Gracefully stops all SmartDoor processes.
+    """
+    process_count = 0
 
-if process_count > 0:
-    print("{0} SmartDoor processes killed".format(process_count))
-else:
-    print('There were no SmartDoor processes running')
+    if MonitorSmartDoor.is_running(False):
+        processes = MonitorSmartDoor.get_processes()
 
-sys.exit()
+        for line in processes.splitlines():
+            process_count += 1
+            line = bytes.decode(line)
+            print("Killing SmartDoor process: {0}".format(line))
+            pid = int(line.split(None, 1)[0])
+            os.kill(pid, signal.SIGTERM)
+
+    if process_count > 0:
+        print("{0} SmartDoor processes killed".format(process_count))
+    else:
+        print('There were no SmartDoor processes running')
+
+    sys.exit()
+
+if __name__ == '__main__':
+    # Being executed as a script
+    kill()
